fix the tests and start implementing the solo class properly
