+++ demos/common_demo_header.hpp
@@ -1,0 +1,60 @@
+/**
+ * @file common_demo_header.hpp
+ * @author Maximilien Naveau (maximilien.naveau@gmail.com)
+ * @brief Contains some default tools for creating demos
+ * @version 0.1
+ * @date 2019-11-07
+ * 
+ * @copyright Copyright (c) 2019
+ * 
+ */
+
+// Exiting on ctrl+c
+#include <signal.h>
+#include <atomic>
+#include <Eigen/Eigen>
+#include <iostream>
+
+#include "real_time_tools/timer.hpp"
+#include "real_time_tools/spinner.hpp"
+#include "real_time_tools/thread.hpp"
+
+/**
+ * @brief This boolean is here to kill cleanly the application upon ctrl+c
+ */
+std::atomic_bool StopDemos(false);
+
+/**
+ * @brief This function is the callback upon a ctrl+c call from the terminal.
+ *
+ * @param s
+ */
+void my_handler(int)
+{
+    StopDemos = true;
+}
+
+/**
+ * @brief Enable to kill the demos cleanly with a ctrl+c
+ */
+void enable_ctrl_c()
+{
+    // make sure we catch the ctrl+c signal to kill the application properly.
+    struct sigaction sigIntHandler;
+    sigIntHandler.sa_handler = my_handler;
+    sigemptyset(&sigIntHandler.sa_mask);
+    sigIntHandler.sa_flags = 0;
+    sigaction(SIGINT, &sigIntHandler, NULL);
+    StopDemos = false;
+}
+
+void print_vector(std::string v_name, Eigen::Ref<Eigen::VectorXd> v)
+{
+    v_name += ": [";
+    rt_printf("%s", v_name.c_str());
+    for (int i = 0; i < v.size(); ++i)
+    {
+        rt_printf("%f, ", v(i));
+    }
+    rt_printf("]\n");
+}
