+++ demos/demo_calibration.cpp
@@ -7,32 +7,13 @@
  * This file uses the Quadruped class in a small demo.
  */
 
-// Exiting on ctrl+c
-#include <signal.h>
-#include <atomic>
 
-#include <cmath>
-#include <deque>
-#include <numeric>
 #include "blmc_robots/blmc_joint_module.hpp"
-#include "real_time_tools/timer.hpp"
+#include "common_demo_header.hpp"
 
-/**
- * @brief This boolean is here to kill cleanly the application upon ctrl+c
- */
-std::atomic_bool StopDemos(false);
-
-/**
- * @brief This function is the callback upon a ctrl+c call from the terminal.
- *
- * @param s
- */
-void my_handler(int s)
-{
-    StopDemos = true;
-}
 
 using namespace blmc_robots;
+
 
 struct Robot
 {
@@ -74,15 +55,9 @@ static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_void_ptr)
     // }
 }  // end control_loop
 
-int main(int argc, char** argv)
+int main(int, char**)
 {
-    // make sure we catch the ctrl+c signal to kill the application properly.
-    struct sigaction sigIntHandler;
-    sigIntHandler.sa_handler = my_handler;
-    sigemptyset(&sigIntHandler.sa_mask);
-    sigIntHandler.sa_flags = 0;
-    sigaction(SIGINT, &sigIntHandler, NULL);
-    StopDemos = false;
+    enable_ctrl_c()
 
     real_time_tools::RealTimeThread thread;
