+++ demos/demo_disentanglement_platform.cpp
@@ -7,15 +7,14 @@
  * This file uses the RealDisentanglementPlatform class in a small demo.
  */
 
-#include <Eigen/Core>
-#include <iostream>
-#include <tuple>
 
+#include <tuple>
 #include "blmc_robots/real_disentanglement_platform.hpp"
-#include "real_time_tools/spinner.hpp"
-#include "real_time_tools/thread.hpp"
+#include "common_demo_header.hpp"
+
 
 using namespace blmc_robots;
+
 
 static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_ptr)
 {
@@ -27,7 +26,7 @@ static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_ptr)
     real_time_tools::Spinner spinner;
     spinner.set_period(0.001);
     size_t count = 0;
-    while (true)
+    while (!StopDemos)
     {
         Eigen::Vector3d desired_torque = 0 * robot->get_max_torques();
 
@@ -46,8 +45,10 @@ static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_ptr)
     }
 }
 
-int main(int argc, char** argv)
+int main(int, char**)
 {
+    enable_ctrl_c();
+
     // set up motor boards -----------------------------------------------------
     auto motor_boards =
         RealDisentanglementPlatform::create_motor_boards("can0", "can1");
