+++ demos/demo_solo_calibration.cpp
@@ -4,11 +4,12 @@
  * \author Maximilien Naveau
  * \date 2018
  *
- * This file uses the Teststand class in a small demo.
+ * This file uses the Solo class in a small demo.
  */
 
 
-#include "blmc_robots/teststand.hpp"
+#include "real_time_tools/timer.hpp"
+#include "blmc_robots/solo.hpp"
 #include "common_demo_header.hpp"
 
 
@@ -17,21 +18,18 @@ using namespace blmc_robots;
 
 static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_void_ptr)
 {
-    Teststand& robot = *(static_cast<Teststand*>(robot_void_ptr));
+    Solo& robot = *(static_cast<Solo*>(robot_void_ptr));
 
     Eigen::Vector2d joint_index_to_zero;
-    joint_index_to_zero[0] = -0.354043;
-    joint_index_to_zero[1] = -0.243433;
+    joint_index_to_zero[0] = 0.0;
+    joint_index_to_zero[1] = 0.0;
+    joint_index_to_zero[2] = 0.0;
+    joint_index_to_zero[3] = 0.0;
     robot.calibrate(joint_index_to_zero);
-
-    long int count = 0;
 
     while(!StopDemos)
     {
-      if(count % 200)
-      {
-        print_vector("Joint Positions", robot.get_joint_positions());
-      }
+      
     }
 
     StopDemos = true;
@@ -45,21 +43,20 @@ int main(int, char**)
 
     real_time_tools::RealTimeThread thread;
 
-    Teststand robot;
+    Solo robot;
 
     robot.initialize();
 
     rt_printf("controller is set up \n");
-    rt_printf("Press enter to launch the calibration \n");
-    char str[256];
-    std::cin.get(str, 256);  // get c-string
 
     thread.create_realtime_thread(&control_loop, &robot);
 
-    // Wait until the application is killed.
-    thread.join();
+    rt_printf("control loop started \n");
 
-    rt_printf("Exit cleanly \n");
+    while (true)
+    {
+        real_time_tools::Timer::sleep_sec(0.01);
+    }
 
     return 0;
 }
