+++ demos/demo_test_bench_8_motors.cpp
@@ -7,28 +7,12 @@
  * This file uses the TestBench8Motors class in a small demo.
  */
 
-#include <signal.h>
-#include <atomic>
+
 #include "blmc_robots/test_bench_8_motors.hpp"
-#include "real_time_tools/spinner.hpp"
-#include "real_time_tools/thread.hpp"
+#include "common_demo_header.hpp"
+
 
 using namespace blmc_robots;
-
-/**
- * @brief This boolean is here to kill cleanly the application upon ctrl+c
- */
-std::atomic_bool StopDemos(false);
-
-/**
- * @brief This function is the callback upon a ctrl+c call from the terminal.
- *
- * @param s
- */
-void my_handler(int s)
-{
-    StopDemos = true;
-}
 
 static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_void_ptr)
 {
@@ -85,15 +69,9 @@ static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_void_ptr)
     return THREAD_FUNCTION_RETURN_VALUE;
 }  // end control_loop
 
-int main(int argc, char** argv)
+int main(int, char**)
 {
-    // make sure we catch the ctrl+c signal to kill the application properly.
-    struct sigaction sigIntHandler;
-    sigIntHandler.sa_handler = my_handler;
-    sigemptyset(&sigIntHandler.sa_mask);
-    sigIntHandler.sa_flags = 0;
-    sigaction(SIGINT, &sigIntHandler, NULL);
-    StopDemos = false;
+    enable_ctrl_c();
 
     TestBench8Motors robot;
