+++ demos/demo_teststand.cpp
@@ -7,44 +7,13 @@
  * This file uses the Teststand class in a small demo.
  */
 
-// Exiting on ctrl+c
-#include <signal.h>
-#include <atomic>
 
-#include <Eigen/Eigen>
-#include <cmath>
-#include <deque>
-#include <numeric>
 #include "blmc_robots/teststand.hpp"
-#include "real_time_tools/timer.hpp"
+#include "common_demo_header.hpp"
+
 
 using namespace blmc_robots;
 
-/**
- * @brief This boolean is here to kill cleanly the application upon ctrl+c
- */
-std::atomic_bool StopDemos(false);
-
-/**
- * @brief This function is the callback upon a ctrl+c call from the terminal.
- *
- * @param s
- */
-void my_handler(int)
-{
-    StopDemos = true;
-}
-
-void print_vector(std::string v_name, Eigen::Ref<Eigen::VectorXd> v)
-{
-    v_name += ": [";
-    rt_printf("%s", v_name.c_str());
-    for (int i = 0; i < v.size(); ++i)
-    {
-        rt_printf("%f, ", v(i));
-    }
-    rt_printf("]\n");
-}
 
 static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_void_ptr)
 {
@@ -144,15 +113,9 @@ static THREAD_FUNCTION_RETURN_TYPE control_loop(void* robot_void_ptr)
     return THREAD_FUNCTION_RETURN_VALUE;
 }  // end control_loop
 
-int main(int argc, char** argv)
+int main(int, char**)
 {
-    // make sure we catch the ctrl+c signal to kill the application properly.
-    struct sigaction sigIntHandler;
-    sigIntHandler.sa_handler = my_handler;
-    sigemptyset(&sigIntHandler.sa_mask);
-    sigIntHandler.sa_flags = 0;
-    sigaction(SIGINT, &sigIntHandler, NULL);
-    StopDemos = false;
+    enable_ctrl_c();
 
     real_time_tools::RealTimeThread thread;
